feat: implement menu primitive
